Added target variable (+1, -1) column to second model
Classified each target column value with +1 (if the value of the target column - Education Level - was over 50%) and -1  (if the value of the target column - Education Level - was under 50%). Like Model 1 (logreg-model1.py) , 0 is assigned for all other values outside this criteria. Prints to terminal.

Minor change of removing a comment from logreg-model1.py.